Add taking photo during rotation fix and refactor edit_parameters directory
