--- app/src/main/java/com/aqchen/filterfiesta/ui/photo_editor/adjustments/edit_parameters/parameter/ParameterEvent.kt
@@ -1,2 +1,0 @@
-package com.aqchen.filterfiesta.ui.photo_editor.adjustments.edit_parameters.parameter
-
