--- app/src/main/java/com/aqchen/filterfiesta/ui/photo_editor/adjustments/edit_parameters/parameter/ParameterViewModel.kt
@@ -1,4 +1,0 @@
-package com.aqchen.filterfiesta.ui.photo_editor.adjustments.edit_parameters.parameter
-
-class ParameterViewModel {
-}
