+++ app/src/main/java/com/aqchen/filterfiesta/ui/photo_editor/edit_parameters/EditParametersPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.aqchen.filterfiesta.ui.photo_editor.adjustments.edit_parameters
+package com.aqchen.filterfiesta.ui.photo_editor.edit_parameters
 
 import android.view.LayoutInflater
 import android.view.View
