+++ app/src/main/java/com/aqchen/filterfiesta/ui/photo_editor/edit_parameters/pager/ParameterPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.aqchen.filterfiesta.ui.photo_editor.adjustments.edit_parameters.pager
+package com.aqchen.filterfiesta.ui.photo_editor.edit_parameters.pager
 
 import android.util.TypedValue
 import android.view.LayoutInflater
