+++ app/src/main/java/com/aqchen/filterfiesta/ui/photo_editor/edit_parameters/pager/ParameterPagerEvent.kt
@@ -1,4 +1,4 @@
-package com.aqchen.filterfiesta.ui.photo_editor.adjustments.edit_parameters.pager
+package com.aqchen.filterfiesta.ui.photo_editor.edit_parameters.pager
 
 sealed class ParameterPagerEvent {
     data class FocusParameterPageViewHolder(val position: Int): ParameterPagerEvent()
