+++ app/src/main/java/com/aqchen/filterfiesta/ui/photo_editor/edit_parameters/parameter/ParameterFragment.kt
@@ -1,4 +1,4 @@
-package com.aqchen.filterfiesta.ui.photo_editor.adjustments.edit_parameters.parameter
+package com.aqchen.filterfiesta.ui.photo_editor.edit_parameters.parameter
 
 import android.os.Bundle
 import android.util.Log
@@ -11,7 +11,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.aqchen.filterfiesta.R
-import com.aqchen.filterfiesta.ui.photo_editor.adjustments.edit_parameters.EditParametersViewModel
+import com.aqchen.filterfiesta.ui.photo_editor.edit_parameters.EditParametersViewModel
 import com.aqchen.filterfiesta.ui.shared_view_models.photo_editor_images.PhotoEditorImagesEvent
 import com.aqchen.filterfiesta.ui.shared_view_models.photo_editor_images.PhotoEditorImagesViewModel
 import com.aqchen.filterfiesta.ui.shared_view_models.photo_editor_images.SelectFilterState
