+++ app/src/main/java/com/aqchen/filterfiesta/ui/photo_editor/edit_parameters/parameter/ParameterViewModel.kt
@@ -1,0 +1,4 @@
+package com.aqchen.filterfiesta.ui.photo_editor.edit_parameters.parameter
+
+class ParameterViewModel {
+}
