+++ app/src/main/res/layout/fragment_bottom_bar_edit_parameters.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.aqchen.filterfiesta.ui.photo_editor.adjustments.edit_parameters.EditParametersFragment"
+    tools:context="com.aqchen.filterfiesta.ui.photo_editor.edit_parameters.EditParametersFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
